+++ Sources/Extensions/StringExtension.swift
@@ -46,6 +46,28 @@ extension String {
         return NSPredicate(format: "SELF MATCHES %@", emailRegex).evaluate(with: self)
     }
 
+    public var isIP4Address: Bool {
+        return confirmIP4isValid(ip4: self)
+    }
+    
+    public var isIP6Address: Bool {
+        return confirmIP6isValid(ip6: self)
+    }
+    
+    public var isIPAddress: Bool {
+        return confirmIP4isValid(ip4: self) || confirmIP6isValid(ip6: self)
+    }
+
+    private func confirmIP4isValid(ip4: String) -> Bool {
+        var sin = sockaddr_in()
+        return ip4.withCString({ cstring in inet_pton(AF_INET, cstring, &sin.sin_addr)}) == 1
+    }
+    
+    private func confirmIP6isValid(ip6: String) -> Bool {
+        var sin6 = sockaddr_in6()
+        return ip6.withCString({ cstring in inet_pton(AF_INET6, cstring, &sin6.sin6_addr)}) == 1
+    }
+
 }
 
 // MARK: - Computed Properties